Exception fix

+++ backend/src/main/java/babel/backend/service/impl/TrashContainerServiceImpl.java
@@ -53,6 +53,9 @@ public class TrashContainerServiceImpl implements TrashContainerService {
     @Override
 
     public void deleteTrashContainer(Long id) {
+        if(trashContainerRepository.findById(id).isEmpty()){
+            throw new ResponseStatusException(HttpStatus.NO_CONTENT, "Trash container with specified id not found.");
+        }
         trashContainerRepository.deleteById(id);
     }
 }
